+++ josh-core/src/cache/distributed.rs
@@ -81,7 +81,7 @@ impl DistributedCacheBackend {
 // The sparse cache is mostly only used for initial "cold starts" or longer "catch up".
 // For incremental filtering it's fine re-filter commits and rely on the local "dense" cache.
 // We store entries for 1% of all commits, and additionally all merges and orphans.
-fn is_eligible(repo: &git2::Repository, oid: git2::Oid, sequence_number: u128) -> bool {
+fn is_eligible(repo: &git2::Repository, oid: git2::Oid, sequence_number: u64) -> bool {
     let parent_count = if let Ok(c) = repo.find_commit(oid) {
         c.parent_ids().count()
     } else {
@@ -94,7 +94,7 @@ fn is_eligible(repo: &git2::Repository, oid: git2::Oid, sequence_number: u128) -
 // To additionally limit the size of the trees the cache is also sharded by sequence
 // number in groups of 10000. Note that this does not limit the number of entries per bucket
 // as branches mean many commits share the same sequence number.
-fn ref_path(key: git2::Oid, sequence_number: u128) -> String {
+fn ref_path(key: git2::Oid, sequence_number: u64) -> String {
     format!(
         "refs/josh/cache/{}/{}/{}",
         CACHE_VERSION,
@@ -116,7 +116,7 @@ impl CacheBackend for DistributedCacheBackend {
         &self,
         filter: Filter,
         from: git2::Oid,
-        sequence_number: u128,
+        sequence_number: u64,
     ) -> anyhow::Result<Option<git2::Oid>> {
         if filter == filter::sequence_number() || filter == filter::reachable_roots() {
             return Ok(None);
@@ -163,7 +163,7 @@ impl CacheBackend for DistributedCacheBackend {
         filter: Filter,
         from: git2::Oid,
         to: git2::Oid,
-        sequence_number: u128,
+        sequence_number: u64,
     ) -> anyhow::Result<()> {
         if filter == filter::sequence_number() || filter == filter::reachable_roots() {
             return Ok(());
